Refactor of repeat payments, additional documentation.
